moved workflow config dropdown under github component
