+++ client-react/src/pages/app/deployment-center/code/DeploymentCenterCodeSettings.tsx
@@ -1,21 +1,21 @@
 import React, { useContext } from 'react';
 import { DeploymentCenterFieldProps, DeploymentCenterCodeFormData } from '../DeploymentCenter.types';
-import DeploymentCenterCodeSource from './DeploymentCenterCodeSource';
 import DeploymentCenterGitHubDataLoader from '../github-provider/DeploymentCenterGitHubDataLoader';
-import { ScmTypes } from '../../../../models/site/config';
+import { ScmType, BuildProvider } from '../../../../models/site/config';
 import DeploymentCenterCodeBuild from './DeploymentCenterCodeBuild';
 import { DeploymentCenterContext } from '../DeploymentCenterContext';
 import DeploymentCenterGitHubReadOnly from '../github-provider/DeploymentCenterGitHubReadOnly';
 import DeploymentCenterCodeBuildReadOnly from './DeploymentCenterCodeBuildReadOnly';
 import DeploymentCenterGitHubWorkflowConfig from '../github-provider/DeploymentCenterGitHubWorkflowConfig';
+import DeploymentCenterCodeSourceAndBuild from './DeploymentCenterCodeSourceAndBuild';
 
 const DeploymentCenterCodeSettings: React.FC<DeploymentCenterFieldProps<DeploymentCenterCodeFormData>> = props => {
   const { formProps } = props;
   const deploymentCenterContext = useContext(DeploymentCenterContext);
 
-  const isGitHubSource = formProps && formProps.values.sourceProvider === ScmTypes.GitHub;
-  const isSourceSelected = formProps && formProps.values.sourceProvider !== ScmTypes.None;
-  const isDeploymentSetup = deploymentCenterContext.siteConfig && deploymentCenterContext.siteConfig.properties.scmType !== ScmTypes.None;
+  const isGitHubSource = formProps && formProps.values.sourceProvider === ScmType.GitHub;
+  const isGitHubActionsBuild = formProps && formProps.values.buildProvider === BuildProvider.GitHubAction;
+  const isDeploymentSetup = deploymentCenterContext.siteConfig && deploymentCenterContext.siteConfig.properties.scmType !== ScmType.None;
 
   const disconnectCallback = () => {
     throw Error('not implemented');
@@ -30,9 +30,9 @@ const DeploymentCenterCodeSettings: React.FC<DeploymentCenterFieldProps<Deployme
         </>
       ) : (
         <>
-          <DeploymentCenterCodeSource />
+          <DeploymentCenterCodeSourceAndBuild formProps={formProps} />
           {isGitHubSource && <DeploymentCenterGitHubDataLoader formProps={formProps} />}
-          {isSourceSelected && <DeploymentCenterCodeBuild formProps={formProps} />}
+          {isGitHubActionsBuild && <DeploymentCenterCodeBuild formProps={formProps} />}
           <DeploymentCenterGitHubWorkflowConfig formProps={formProps} />
         </>
       )}
