+++ client-react/src/pages/app/deployment-center/container/DeploymentCenterContainerSettings.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import DeploymentCenterContainerSource from './DeploymentCenterContainerSource';
 import { ContainerRegistrySources, DeploymentCenterFieldProps, DeploymentCenterContainerFormData } from '../DeploymentCenter.types';
-import { ScmTypes } from '../../../../models/site/config';
+import { ScmType } from '../../../../models/site/config';
 import DeploymentCenterContainerRegistrySettings from './DeploymentCenterContainerRegistrySettings';
 import DeploymentCenterContainerAcrSettings from './DeploymentCenterContainerAcrSettings';
 import DeploymentCenterContainerDockerHubSettings from './DeploymentCenterContainerDockeHubSettings';
@@ -10,7 +10,7 @@ import DeploymentCenterGitHubDataLoader from '../github-provider/DeploymentCente
 
 const DeploymentCenterContainerSettings: React.FC<DeploymentCenterFieldProps<DeploymentCenterContainerFormData>> = props => {
   const { formProps } = props;
-  const isGitHubActionEnabled = formProps && formProps.values.scmType === ScmTypes.GitHubAction;
+  const isGitHubActionEnabled = formProps && formProps.values.scmType === ScmType.GitHubAction;
   const isAcrConfigured = formProps && formProps.values.registrySource === ContainerRegistrySources.acr;
   const isDockerHubConfigured = formProps && formProps.values.registrySource === ContainerRegistrySources.docker;
   const isPrivateRegistryConfigured = formProps && formProps.values.registrySource === ContainerRegistrySources.privateRegistry;
